refactor(domain): lottoPrizeMoney long형을 변경
LottoPrize
- 1등 2명만 나와도 int형을 벗어나기 때문에, long형으로 변

LottoResults
- LottoPrize 내 형변환 진행함에 따라 LottoResults도 변경

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -13,9 +13,9 @@ public enum LottoPrize {
     ;
 
     private final int matchCount;
-    private final int lottoPrizeMoney;
+    private final long lottoPrizeMoney;
 
-    LottoPrize(int matchCount, int lottoPrizeMoney) {
+    LottoPrize(int matchCount, long lottoPrizeMoney) {
         this.matchCount = matchCount;
         this.lottoPrizeMoney = lottoPrizeMoney;
     }
@@ -54,7 +54,7 @@ public enum LottoPrize {
         return matchCount;
     }
 
-    public int getLottoPrizeMoney() {
+    public long getLottoPrizeMoney() {
         return lottoPrizeMoney;
     }
 }
